+++ internal/cli/base_init.go
@@ -8,7 +8,7 @@ import (
 	"github.com/hashicorp/waypoint/internal/clicontext"
 	clientpkg "github.com/hashicorp/waypoint/internal/client"
 	configpkg "github.com/hashicorp/waypoint/internal/config"
-	"github.com/hashicorp/waypoint/internal/serverclient"
+	"github.com/hashicorp/waypoint/pkg/serverclient"
 )
 
 // This file contains the various methods that are used to perform
