+++ internal/cli/runner_agent.go
@@ -17,8 +17,8 @@ import (
 	"github.com/hashicorp/waypoint/internal/pkg/flag"
 	"github.com/hashicorp/waypoint/internal/plugin"
 	runnerpkg "github.com/hashicorp/waypoint/internal/runner"
-	"github.com/hashicorp/waypoint/internal/serverclient"
 	pb "github.com/hashicorp/waypoint/pkg/server/gen"
+	"github.com/hashicorp/waypoint/pkg/serverclient"
 )
 
 type RunnerAgentCommand struct {
